refactor: Review 엔터티의 grade 필드의 타입을 int로 변경
- double -> int

+++ src/main/java/wad/seoul_nolgoat/web/review/dto/request/ReviewUpdateDto.java
@@ -7,6 +7,6 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ReviewUpdateDto {
 
-    private final double grade;
+    private final int grade;
     private final String content;
 }
