+++ src/main/java/wad/seoul_nolgoat/web/review/dto/response/ReviewDetailsForUserDto.java
@@ -7,7 +7,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ReviewDetailsForUserDto {
 
-    private final double grade;
+    private final int grade;
     private final String content;
     private final String imageUrl;
     private final Long storeId;
